trying to fix the cache error problem

+++ lib/features/Home/data/model/request_model.dart
@@ -155,7 +155,7 @@ class MenuItems {
   SizesPrices? sizesPrices;
   bool? healthyMode;
   String? description;
-  int? rate;
+  double? rate;
   String? photo;
   dynamic restaurantPhoto;
   dynamic restaurantName;
